Task recipes-base-auth completed
