+++ components/Templates/Recipes/RecipeTemplate/RecipeTemplate.module.scss
@@ -11,6 +11,10 @@
     grid-template-columns: 1fr;
   }
 
+  &__author {
+    margin-top: 10px;
+  }
+
   &__header {
     grid-column: 1/-1;
   }
